Geo service is done

+++ src/main/java/com/khorunaliyev/kettu/dto/projection/RegionInfo.java
@@ -11,7 +11,7 @@ import java.util.Set;
  */
 public interface RegionInfo {
 
-    Long getId();
+    Integer getId();
 
     String getName();
